fixed xtokens as per suggeestion; unified trait impl for statit route
