+++ code/parachain/runtime/common/src/topology.rs
@@ -28,5 +28,16 @@ pub mod relay {
 	pub const LOCATION: MultiLocation = MultiLocation { parents: 1, interior: Here };
 }
 
-use xcm::latest::prelude::*;
-pub const SELF_RECURSIVE: MultiLocation = MultiLocation { parents: 0, interior: Here };
+pub mod this {
+	use frame_support::parameter_types;
+	use xcm::latest::prelude::*;
+	pub const LOCAL: MultiLocation = MultiLocation { parents: 0, interior: Here };
+	parameter_types! {	
+		pub const Local: MultiLocation = LOCAL;
+	}
+
+	pub fn sibling(para_id: u32) -> MultiLocation{
+		MultiLocation::new(1, X1(Parachain(para_id)))
+	}
+	
+}
